% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{dependency-class}
\alias{dependency}
\alias{dependency-class}
\title{Dependency class}
\description{
Dependency class
}
\section{Methods}{

\describe{
\item{\code{update(df.update)}}{Updates the dependency object with a read in or write out}
}}

