% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.project.dependency.db}
\alias{get.project.dependency.db}
\title{Get project dependency database}
\usage{
get.project.dependency.db(project.id)
}
\arguments{
\item{project.id}{}
}
\value{
data.frame dump of database
}
\description{
Get project dependency database
}
\details{
Assumes project.id is the name of a database.
}

