% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.project.info.si}
\alias{get.project.info.si}
\title{Given source_info object, retrieves project information}
\usage{
get.project.info.si(source_info)
}
\arguments{
\item{source_info}{is list with source information}
}
\value{
list with stacked dependency files, graph of dependencies, and condensed file information
}
\description{
Given source_info object, retrieves project information
}

