% Generated by roxygen2 (4.0.1): do not edit by hand
\name{git.info}
\alias{git.info}
\title{Retrieves the information from git about a file}
\usage{
git.info(gitdir, filename, branch = NULL, git_args = character(),
  git_binary = NULL)
}
\arguments{
\item{gitdir}{string with git directory}

\item{filename}{string of file to query}

\item{branch}{git branch}

\item{git_args}{string argument for git}

\item{git_binary}{location of git executable}
}
\value{
git log for filename
}
\description{
Retrieves the information from git about a file
}

