% Generated by roxygen2 (4.0.1): do not edit by hand
\name{graft.branch}
\alias{graft.branch}
\title{Copy dependent programs from swap directory into Program directory}
\usage{
graft.branch(branch_name, run = TRUE, start.up = FALSE, project.id = NULL,
  overwriteTF = FALSE)
}
\arguments{
\item{branch_name}{Name of directory containing branch data to graft}

\item{run}{logical Run scripts or not}

\item{project.id}{Name of project to locate swap and project home directory}

\item{start.up}{= TRUE create program directory}

\item{overwriteTF}{logical Overwrite existing programs or not}
}
\value{
Branch information
}
\description{
Copy dependent programs from swap directory into Program directory
}
\details{
Runs as test and to create dependency files
}

