% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Harvest.trees}
\alias{Harvest.trees}
\title{Collect trees from dependency directory}
\usage{
Harvest.trees(dependency.dir)
}
\arguments{
\item{dependency.dir}{Directory with dependency files}
}
\value{
data frame of stacked dependency files
}
\description{
Collect trees from dependency directory
}

