% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intialize.dependency.info.R
\name{initialize.dependency.info}
\alias{initialize.dependency.info}
\title{Initializes dependency object source_info}
\usage{
initialize.dependency.info(source_info_arg)
}
\arguments{
\item{source_info_arg}{is a source_info list with describing R script and project}
}
\value{
Depedency file location
}
\description{
Initializes dependency object source_info
}

