% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Make.dependency.graph.obj}
\alias{Make.dependency.graph.obj}
\title{Creates an graph object from a dependency object}
\usage{
Make.dependency.graph.obj(dependency.out)
}
\arguments{
\item{dependency.out}{Dependency object(s) to make graph out of}
}
\value{
graph object of project/program dependencies
}
\description{
Creates an graph object from a dependency object
}

