% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.hyperlink}
\alias{make.hyperlink}
\title{Makes HTML hyper link}
\usage{
make.hyperlink(files, links)
}
\arguments{
\item{files}{character vector of filenames}
}
\value{
link command vector
}
\description{
Makes HTML hyper link
}
\details{
Used in making HTML files
}

