% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Make.summary.graph.si}
\alias{Make.summary.graph.si}
\title{Returns summary graph of dependencies given source_info object}
\usage{
Make.summary.graph.si(source_info, plot.graph = FALSE)
}
\arguments{
\item{source_info}{Source information list}

\item{plot.graph}{logical indicating requet to make plot}
}
\value{
graph of dependency relationships
}
\description{
Returns summary graph of dependencies given source_info object
}

