% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Match.file.hash}
\alias{Match.file.hash}
\title{Matches hash of target file to target file in dependency tree}
\usage{
Match.file.hash(tree, file.for.hashing)
}
\arguments{
\item{tree}{Dependency data frame}

\item{file.for.hashing}{In filesystem to compare hash with dependency tree}
}
\value{
Returns rows of tree that match file.for.hasing
}
\description{
Matches hash of target file to target file in dependency tree
}
\details{
Only for result files.
}

