% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mystar.R
\name{mystar}
\alias{mystar}
\title{Star function for igraph plotting}
\usage{
mystar(coords, v = NULL, params)
}
\arguments{
\item{coords}{coordinates for plotting}

\item{v}{vertex}

\item{params}{graphics parameters}
}
\value{
nothing returned
}
\description{
Star function for igraph plotting
}

