% Generated by roxygen2 (4.0.1): do not edit by hand
\name{path.expand.2}
\alias{path.expand.2}
\title{Swap / for \\ in path expand}
\usage{
path.expand.2(x)
}
\arguments{
\item{x}{file path, could be relative or ~}
}
\value{
Full filepath to x
}
\description{
Swap / for \\ in path expand
}
\details{
See path.expand() in base R
}

