% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swap.project.paths}
\alias{swap.project.paths}
\title{Take list of dependency file data and changes the project path}
\usage{
swap.project.paths(list.deps,
  new.path = get.project.path(source_info$project.id))
}
\arguments{
\item{list.deps}{list of dependency file data}

\item{new.path}{file path for the new project path}
}
\value{
Updated list of dependency data
}
\description{
Take list of dependency file data and changes the project path
}
\details{
Used with swapping branches by rework.project.path()
}

