% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Sync.test}
\alias{Sync.test}
\title{check the synchrony of source files and their created objects}
\usage{
Sync.test(dagger, tree, plotl = TRUE)
}
\arguments{
\item{g.all}{dependency object or set of dependency objects making a directed acyclic graph}

\item{tree}{dependency tree corresponding to dagger}

\item{plotl}{logical for plotting or not}
}
\description{
check the synchrony of source files and their created objects
}

