% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Write}
\alias{Write}
\title{Write object and capture file information}
\usage{
Write(obj = NULL, file.name = "data.csv", description = "Result file",
  write.fcn = guess.write.fcn(file.name), ...)
}
\arguments{
\item{file.name}{file to write to the source "Result" directory}

\item{description}{describes object to write}

\item{write.fcn}{function for writing file of object type}
}
\value{
File information list
}
\description{
Write object and capture file information
}

