% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mystar}
\alias{mystar}
\title{Star function for igraph plotting}
\usage{
mystar(coords, v = NULL, params)
}
\arguments{
\item{coords}{coordinates for plotting}

\item{v}{vertex}

\item{param}{graphics parameters}
}
\description{
Star function for igraph plotting
}

